# Conflicts:
#	src/main/java/gui/collections/CollectionWindowController.java

+++ src/main/java/common/Human.java
@@ -261,8 +261,8 @@ public class Human implements Comparable<Human> {
         return status;
     }  
     public void setRandomCords(){
-        this.coordinates.x = new Random().nextDouble(500);
-        this.coordinates.y = new Random().nextDouble(500);
+        this.coordinates.x = (double) new Random().nextInt(500);
+        this.coordinates.y = (double) new Random().nextInt(500);
     }
     public String getStats(){
         return statsToString();
